requirements.txt
annotated-types==0.7.0
black==25.1.0
click==8.1.8
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pytest==8.3.5
PyYAML==6.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
